refactor(web-integration): split CDP-mode helpers out of mcp-tools-cdp
`mcp-tools-cdp.ts` had grown to handle six concerns at once: CDP URL
resolution, proxy process lifecycle, proxy metadata IO, cross-command
targetId storage, puppeteer agent construction and the MCP tool schema.
Split the proxy-side and store-side concerns into dedicated modules so
the file is just the tool class plus its puppeteer glue.

- New `cdp-proxy-manager.ts` owns the proxy lifecycle and its on-disk
  metadata: `isProxyAlive`, `readProxyEndpoint`, `readProxyUpstream`,
  `killProxy`, `getProxyEndpoint`, plus the page-level URL resolver.
  All of these are now real module exports rather than file-private
  helpers behind a `__test__` namespace.
- New `cdp-target-store.ts` owns the cross-command targetId persistence
  (`readSavedTargetId`, `saveTargetId`, `cleanupTargetIdFile`).
- `mcp-tools-cdp.ts` keeps `WebCdpMidsceneTools`, the puppeteer
  `getTargetId()` glue and the target-discovery delay constant. Its
  external surface (`WebCdpMidsceneTools`) is unchanged.
- Tests now import from `cdp-proxy-manager.js` directly; the
  `__test__` namespace is gone.
